pyproject.toml
[project]
name = "guard"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard] (>=0.124.0,<0.125.0)",
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "alembic-postgresql-enum (>=1.8.0,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "aiomysql (>=0.3.2,<0.4.0)",
    "mypy (>=1.19.0,<2.0.0)",
    "ruff (>=0.14.8,<0.15.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "apiexception (>=0.2.1,<0.3.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)"
]

[tool.poetry]
packages = [{include = "guard", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
extend-select = ["I", "TRY", "UP"]
per-file-ignores = {"tests/*" = ["TRY"], "alembic/*" = ["F401"]}



